vm: implement the ctrl-break handler (#1573)
The code interrupts busy loops by making the main thread throw the
STATUS_SINGLE_STEP exception. This does not start the low-level debugger
mode in the console, but rather returns control to the Listener, which is
very nice and useful. When the Listener is not in a tight loop, the code
does nothing (probably the exception is suppressed somewhere by the
Listener).

Unfortunately, sometimes the whole process crashes. This seems to happen
when the GC is trapped. It would be good to be able to only interrupt the
main thread when it is running the user code by analyzing the Eip value so
that VM code would be exempt from trapping. If that is not feasible, at
least we can make the GC provide some sort of signal to the Ctrl-Break
handler that it's not a good time for an interruption.

Another problem with the code is that it doesn't always work, you need to
activate another window (but not the Factor's console window!) to make it
react to Ctrl-Break. This is an issue with GetAsyncKeyState and the way
Windows handles privilege separation. For instance, if Factor is running
as a non-admin user, and you switch focus to another app with admin
privileges and press Ctrl-Break, then Factor's thread won't detect that.
If you do the same thing, but with the other process having the same
rights as Factor, it will handle the Ctrl-Break event. The funny thing is,
if the Listener hangs, its main window is replaced with a more responsive
dummy, which is owned by the Explorer.exe, and that substitution prevents
Ctrl-Break from working on systems newer than XP.

Yet another issue is that the keystroke is sensed globally, so if you have
multiple Factor instances running, they will all get the signal. The good
news is that only the ones in the tight loop seem to react.