vm: use enqueue_fep to signal ctrl-break
The problem of interrupting the app when it's not hanging is not
solved. When Factor is in a tight loop, the chances of a successful
interruption are very high, but if it works fine, then pressing
Ctrl-Break simply crashes the it. It would be much better if we were
able to ignore the signal.