!TODO! Review this commit
This commit was done before the fix by KUSUMOTO Norio was accepted, so I
have to make sure the rest of the changes here are relevant (see original
comments below). If not, I may prefer the fixes made in my-char-fixes and
integrate only a part of Kusumoto's changes, or something. If yes, I will
have to propose my changes to the Factor upstream code.

ui.backend.windows: fix special treatment of the character keypresses

Without this fix, Caps Lock state interferes with the shortcuts like
Ctrl-a by registering them as Ctrl-A, and it is also impossible to create
shortcuts with the Shift key to be triggered only by the physical pressing
of the Shift key, and not the Caps Lock state.

With this fix, all the key-down events for characters have the lower case
letter, and the Shift state reported in the mods field.